refactor: 라벨 주소 변경
