+++ src/components/modal/RequestSupportModal.tsx
@@ -4,7 +4,7 @@ import ArrowIcon from '../../assets/icons/arrowUp2Thin.svg?react';
 import CheckLightIcon from '../../assets/icons/checkLight.svg?react';
 import ApplyIcon from '../../assets/icons/apply.svg?react';
 import UnapplyIcon from '../../assets/icons/unapply.svg?react';
-import { getCategoryLabel } from '../../constants/categoryMapper';
+import { getCategoryLabel } from '../../utils/categoryMapping';
 import { useNavigate } from 'react-router-dom';
 
 import { useMutation } from '@tanstack/react-query';
